+++ R/copy-to.r
@@ -7,10 +7,14 @@
 #' @param dest remote data source
 #' @param df local data frame
 #' @param name name for new remote table.
+#' @param overwrite If `TRUE`, will overwrite an existing table with
+#'   name `name`. If `FALSE`, will throw an error if `name` already
+#'   exists.
 #' @param ... other parameters passed to methods.
 #' @return a `tbl` object in the remote source
 #' @export
-copy_to <- function(dest, df, name = deparse(substitute(df)), ...) {
+copy_to <- function(dest, df, name = deparse(substitute(df)),
+                    overwrite = FALSE, ...) {
   UseMethod("copy_to")
 }